Refactor step arguments to own module
