+++ robotmbt/suiteprocessors.py
@@ -39,6 +39,7 @@ from robot.api import logger
 
 from .suitedata import Suite, Scenario, Step
 from .tracestate import TraceState
+from .steparguments import StepArguments
 
 class SuiteProcessors:
     @not_keyword
@@ -255,12 +256,12 @@ class SuiteProcessors:
                         edge_step = Step('Log', f"Refinement follows for step: {step.keyword}", parent=scenario)
                         edge_step.gherkin_kw = step.gherkin_kw
                         edge_step.model_info = dict(IN=step.model_info['IN'], OUT=[])
-                        edge_step.emb_args = step.emb_args[:]
+                        edge_step.emb_args = StepArguments(step.emb_args)
                         front.steps.append(edge_step)
                         edge_step = Step('Log', f"Refinement completed for step: {step.keyword}", parent=scenario)
                         edge_step.gherkin_kw = step.gherkin_kw
                         edge_step.model_info = dict(IN=[], OUT=step.model_info['OUT'])
-                        edge_step.emb_args = step.emb_args[:]
+                        edge_step.emb_args = StepArguments(step.emb_args)
                         back.steps[0] = copy.deepcopy(back.steps[0])
                         back.steps[0].model_info = dict(IN=[], OUT=[])
                         back.steps.insert(0, edge_step)
@@ -285,7 +286,7 @@ class SuiteProcessors:
             for expr in SuiteProcessors._relevant_expressions(step):
                 try:
                     if m.process_expression(expr, step.emb_args) is False:
-                        logger.debug(f"Unable to insert scenario {scenario.name} due to step {step.keyword}: {expr} is False") # report with filled in args
+                        logger.debug(f"Unable to insert scenario {scenario.name} due to step {step.keyword}: {expr} is False")
                         logger.debug(f"last state:\n{m.get_status_text()}")
                         return False
                 except Exception as err:
